Relax OwnedByService return value restriction
Allow C API functions to return handles to objects with OwnedByService
lifetime semantics. I.e. handles to objects that already exist in the
service and whose lifetime is tied only to the service.

This is a little dangerous, since an object may be destroyed in the
service while the client still holds a handle, but is still a useful
workflow for many scenarios.