Remove duplication + add `free_fn_ptr` wrapper
We must have the decorated function available at compile-time or
`static` in order to create decorating lambdas that do not need to
capture (and so can be converted to C suite function poitners).

Previously, accepting a function pointer as an argument to `decorate`
was dangerous, because the function was stored in a `static` "cache"
that is meant to be unique for the function decorated, but may not be in
the case of function pointers, since they are cached (template-matched)
on signature, not identity.

So add a `free_fn_ptr` wrapper, in a similar vein to `mem_fn_ptr` to
allow compile time passing of function pointer, removing the need for
a `static` cache of the function.

Also reduce duplication by rolling implementation into a single
`decorate_impl` function.

Hence refactor lambda handling to avoid the need for a `static` cache
at all, by using dodgy `reinterpret_cast` trick inside a `static`
method.